# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Outfit Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ClothingItemCategory = typeof ClothingItemCategory[keyof typeof ClothingItemCategory];


export const ClothingItemCategory = {
  tops: 'tops',
  bottoms: 'bottoms',
  shoes: 'shoes',
  accessories: 'accessories',
  outerwear: 'outerwear',
  dresses: 'dresses',
} as const;

export interface ClothingItem {
  id: number;
  name: string;
  category: ClothingItemCategory;
  /** @nullable */
  imageObjectPath: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  occasion?: string | null;
  /** @nullable */
  purchasePrice?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  notes?: string | null;
  isFavorite?: boolean;
  timesWorn?: number;
  createdAt: string;
}

export type ClothingItemInputCategory = typeof ClothingItemInputCategory[keyof typeof ClothingItemInputCategory];


export const ClothingItemInputCategory = {
  tops: 'tops',
  bottoms: 'bottoms',
  shoes: 'shoes',
  accessories: 'accessories',
  outerwear: 'outerwear',
  dresses: 'dresses',
} as const;

export interface ClothingItemInput {
  /** @minLength 1 */
  name: string;
  category: ClothingItemInputCategory;
  imageObjectPath?: string;
  color?: string;
  brand?: string;
  size?: string;
  season?: string;
  occasion?: string;
  purchasePrice?: string;
  purchaseDate?: string;
  notes?: string;
  isFavorite?: boolean;
}

export type ClothingItemUpdateCategory = typeof ClothingItemUpdateCategory[keyof typeof ClothingItemUpdateCategory];


export const ClothingItemUpdateCategory = {
  tops: 'tops',
  bottoms: 'bottoms',
  shoes: 'shoes',
  accessories: 'accessories',
  outerwear: 'outerwear',
  dresses: 'dresses',
} as const;

export interface ClothingItemUpdate {
  /** @minLength 1 */
  name?: string;
  category?: ClothingItemUpdateCategory;
  imageObjectPath?: string;
  color?: string;
  brand?: string;
  size?: string;
  season?: string;
  occasion?: string;
  purchasePrice?: string;
  purchaseDate?: string;
  notes?: string;
  isFavorite?: boolean;
  timesWorn?: number;
}

export interface CategoryStat {
  category: string;
  count: number;
}

export interface WardrobeStats {
  total: number;
  byCategory: CategoryStat[];
  favorites: number;
  outfitsGenerated: number;
}

export interface GenerateOutfitOptions {
  excludeCategories?: string[];
}

export interface GeneratedOutfit {
  items: ClothingItem[];
}

export interface SavedOutfit {
  id: number;
  name: string;
  /** @nullable */
  notes?: string | null;
  itemIds: number[];
  items?: ClothingItem[];
  createdAt: string;
}

export interface OutfitInput {
  /** @minLength 1 */
  name: string;
  notes?: string;
  /** @minItems 1 */
  itemIds: number[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type ListClothingParams = {
/**
 * Filter by category
 */
category?: ListClothingCategory;
};

export type ListClothingCategory = typeof ListClothingCategory[keyof typeof ListClothingCategory];


export const ListClothingCategory = {
  tops: 'tops',
  bottoms: 'bottoms',
  shoes: 'shoes',
  accessories: 'accessories',
  outerwear: 'outerwear',
  dresses: 'dresses',
} as const;