# lib/api-zod/src/generated/types/clothingItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Outfit Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ClothingItemCategory } from './clothingItemCategory';

export interface ClothingItem {
  id: number;
  name: string;
  category: ClothingItemCategory;
  /** @nullable */
  imageObjectPath: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  occasion?: string | null;
  /** @nullable */
  purchasePrice?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  notes?: string | null;
  isFavorite?: boolean;
  timesWorn?: number;
  createdAt: Date;
}