# lib/api-zod/src/generated/types/clothingItemUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Outfit Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ClothingItemUpdateCategory } from './clothingItemUpdateCategory';

export interface ClothingItemUpdate {
  /** @minLength 1 */
  name?: string;
  category?: ClothingItemUpdateCategory;
  imageObjectPath?: string;
  color?: string;
  brand?: string;
  size?: string;
  season?: string;
  occasion?: string;
  purchasePrice?: string;
  purchaseDate?: string;
  notes?: string;
  isFavorite?: boolean;
  timesWorn?: number;
}